First working version: integrated image-upload-to-iiif workflow
