+++ src/backend/helpers/documentHelpers.ts
@@ -6,12 +6,61 @@ import type { Response } from '@backend/Types';
 import type { Document, DocumentInContext, Layer } from 'src/Types';
 
 /**
- * Initializes a new Document in a Context.
+ * Initializes a new Document in a Context. Process differs for
+ * different types of content.
  * 
- * 1. creates the Document record.
- * 2. creates a default Layer on the document in the given Context.
+ * - For IIIF remote sources, only the URL of the manifest is stored in Supabase
+ * - Text and TEI documents are uploaded to Supabase storage
+ * - Images are uploaded to the IIIF server, and a reference to the image
+ *   manifest stored in Supabase
  */
 export const initDocument = (
+  supabase: SupabaseClient, 
+  name: string, 
+  projectId: string, 
+  contextId: string,
+  onProgress?: (progress: number) => void,
+  file?: File,
+  url?: string
+): Promise<DocumentInContext> => { 
+  if (file?.type.startsWith('image')) {
+    console.log
+    // If the document is an image upload, the file is first
+    // uploaded to the IIIF server, and then treated like a remote
+    // IIIF source.
+    return uploadImage(supabase, file, name, onProgress)
+      .then(iiif => 
+          _initDocument(
+            supabase, 
+            name, 
+            projectId, 
+            contextId, 
+            undefined, 
+            undefined,
+            iiif.manifest_iiif_url));
+  } else {
+    return _initDocument(
+      supabase, 
+      name, 
+      projectId, 
+      contextId, 
+      onProgress, 
+      file, 
+      url);
+  }
+}
+
+/**
+ * Initializes a text (plaintext, TEI) or remote IIIF document.
+ * 
+ * Procedure is as follows:
+ * 1. A `document` record is created in Supabase.
+ * 2. A default `layer` on the document is created in the given context. (Happens in parallel.)
+ * 3. After step 1 and 2 have completed: 
+ *    - the method returns if the document is a remote IIIF source
+ *    - the text file is uploaded to Supabase storage, with a reference to the document ID
+ */
+const _initDocument = (
   supabase: SupabaseClient, 
   name: string, 
   projectId: string, 
@@ -44,16 +93,6 @@ export const initDocument = (
       }
     })
 }
-
-/** Just temporary... */
-export const initImageDocument = (
-  supabase: SupabaseClient, 
-  name: string, 
-  projectId: string, 
-  contextId: string,
-  file: File,
-  onProgress?: (progress: number) => void,
-) => uploadImage(supabase, file, name, onProgress);
 
 export const listDocumentsInProject = (
   supabase: SupabaseClient,
